Return tree base path and files found.
In order to confirm that the contents of a trees output are correct you
must be able to `fs.readFileSync(path.join(baseDir, itemPath))`. Without
this change, we only have the relative path for each output file so we
can confirm that files exist in the correct location, but not assert
that their content is correct.